Move pagination and sorting tests into own file
